[300 line limit] Refactor CohortEditor, OverallMetricChart, set max-lines rule for core-ui package (#1612)
* refactor AxisConfigDialog

* reorder props

* refactor OverallMetricChart

* refactor CohortEditor and CohortEditorFilter

* refactor JointDataset

* update core-ui eslint rules

* pass this.state.cohortName instead of this.props.cohortName to CohortEditorPanelContent

* update import and remove unused state
