+++ libs/core-ui/.eslintrc.json
@@ -1,6 +1,6 @@
 {
   "extends": ["../../.eslintrc"],
-  "ignorePatterns": ["!**/*", "Null.tsx"],
+  "ignorePatterns": ["!**/*", "Null.tsx", "JointDataset.ts"],
   "overrides": [
     {
       "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
@@ -16,7 +16,9 @@
     },
     {
       "files": ["*.ts", "*.tsx"],
-      "rules": {}
+      "rules": {
+        "max-lines": "error"
+      }
     },
     {
       "files": ["*.js", "*.jsx"],
